requirements.txt
langchain>=0.1.0
langchain-openai
langchain-core
elevenlabs
openai
pyaudio
numpy
python-dotenv
langchain-google-genai
